+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 
 ## [Unreleased]
 
-TODO: add at least one Added, Changed, Deprecated, Removed, Fixed or Security section
+- add new GUI elements: Table, TableHead, TableBody, TableRow, TableCell
 
 ## [2.31.4] 2018-06-13