+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 
 ### Added
 - add new GUI elements: Table, TableHead, TableBody, TableRow, TableCell
+- add `hideLabel` option to `<Checkbox/>` and `<Radio/>` elements, when true then labels are not visible but accessible
+
 ### Changed
 - Pagination: left default alignement and smaller font sizes