+++ CHANGELOG.md
@@ -12,6 +12,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 - Pagination: left default alignement and smaller font sizes
 - Data Table: styles now provided by new `<Table/>` element
 
+### Removed
+- remove support for `$color-primary-dark`, `$ecc-typo-display-3-font-size`, `$ecc-typo-display-2-font-size`, `$ecc-typo-display-1-font-size`, `$ecc-typo-headline-font-size`, `$ecc-typo-title-font-size`, `$ecc-typo-subhead-font-size` and remove the info about their deprecation
+
 ## [2.31.4] 2018-06-13
 
 ### Added